# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a captcha service API for Skystrike.",
        "title": "Skystrike Captcha API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/captcha/generate": {
            "get": {
                "description": "Generates a new base64 captcha image and a corresponding token",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "Generate Captcha",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/captcha/verify": {
            "post": {
                "description": "Verifies if the provided captcha text matches the token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "Verify Captcha",
                "parameters": [
                    {
                        "description": "Captcha verification payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/captcha.CaptchaVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "captcha.CaptchaVerifyRequest": {
            "type": "object",
            "required": [
                "text",
                "token"
            ],
            "properties": {
                "text": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        }
    }
}